Refactor reweigh updatePaymentRequest template
