+++ src/constants/MoveHistory/EventTemplates/UpdateReweigh/updatePaymentRequest.test.jsx
@@ -1,9 +1,10 @@
+import { render, screen } from '@testing-library/react';
+
 import getTemplate from 'constants/MoveHistory/TemplateManager';
 import o from 'constants/MoveHistory/UIDisplay/Operations';
-import e from 'constants/MoveHistory/EventTemplates/reweighPaymentRequest';
 
 describe('reweighs update', () => {
-  const item = {
+  const historyRecord = {
     action: 'UPDATE',
     eventName: o.updateReweigh,
     tableName: 'payment_requests',
@@ -12,8 +13,10 @@ describe('reweighs update', () => {
     oldValues: { payment_request_number: '0288-7994-1' },
   };
   it('correctly matches the reweigh payment request', () => {
-    const result = getTemplate(item);
-    expect(result).toMatchObject(e);
-    expect(result.getStatusDetails(item)).toEqual('Recalculated payment request');
+    const template = getTemplate(historyRecord);
+
+    render(template.getDetails(historyRecord));
+    expect(screen.getByText('Status')).toBeInTheDocument();
+    expect(screen.getByText(': Recalculated payment request')).toBeInTheDocument();
   });
 });
